+++ spec/errors/missing_auth_spec.rb
@@ -1,10 +1,10 @@
 require 'spec_helper'
-require 'yt/errors/missing_auth'
+require 'yt/errors/unauthorized'
 
-describe Yt::Errors::MissingAuth do
+describe Yt::Errors::Unauthorized do
   let(:msg) { %r{^A request to YouTube API was sent without a valid authentication} }
 
   describe '#exception' do
-    it { expect{raise Yt::Errors::MissingAuth}.to raise_error msg }
+    it { expect{raise Yt::Errors::Unauthorized}.to raise_error msg }
   end
 end